使用DrawView绘制TableViewCell的主要控件，达到优化tableView性能！ 添加图片浏览器功能！

+++ WUO/WUO/Util/Refresh/XYRefreshGifHeader.m
@@ -18,8 +18,8 @@
         self.stateLabel.hidden = YES;
         
         NSMutableArray *images = [NSMutableArray arrayWithCapacity:16];
-        for (NSInteger i = 1; i < 17; i++) {
-            UIImage *image = [UIImage imageNamed:[NSString stringWithFormat:@"refresh_%ld", i]];
+        for (int i = 1; i < 17; i++) {
+            UIImage *image = [UIImage imageNamed:[NSString stringWithFormat:@"refresh_%d", i]];
             [images addObject:image];
         }
